refactor: admin에서 공통으로 사용할 미리보기 dialog 추출
